Update install.md
So just wanted to let you know that the original links on this page weren't working, going to them gave you a blank page with the words: "Couldn't find the requested release version 3.5.2". I updated with url's that worked that I found on cdnjs. Feel free to disregard this change if there is another cdn that's more appropriate here.